requirement.txt
absl-py==2.3.0
annotated-types==0.7.0
anyio==3.7.1
attrs==25.3.0
cffi==1.17.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
exceptiongroup==1.3.0
fastapi==0.115.12
filelock==3.18.0
flatbuffers==25.2.10
fonttools==4.58.1
fsspec==2025.5.1
h11==0.16.0
idna==3.10
importlib-metadata==8.7.0
importlib-resources==6.5.2
jax==0.4.30
jaxlib==0.4.30
jinja2==3.1.6
joblib==1.4.2
kiwisolver==1.4.7
MarkupSafe==3.0.2
matplotlib==3.9.4
mediapipe==0.10.21
ml-dtypes==0.5.1
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.4
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opt-einsum==3.4.0
packaging==25.0
pandas==2.2.3
pillow==11.2.1
praat-parselmouth==0.4.5
protobuf==4.25.8
pycparser==2.22
pydantic==2.6.4
pydantic-core==2.16.3
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
scikit-learn==1.5.1
scipy==1.13.1
sentencepiece==0.2.0
six==1.17.0
sniffio==1.3.0
sounddevice==0.5.2
starlette==0.46.2
sympy==1.14.0
threadpoolctl==3.6.0
torch==2.7.0
typing-extensions==4.13.2
typing-inspection==0.4.1
tzdata==2025.2
uvicorn==0.34.2
watchfiles==0.15.0
websockets==12.0
zipp==3.22.0